Feat : set check for the ID if null or not in the getById and update && delete

+++ src/main/java/net/axel/services/implementations/BaseService.java
@@ -25,6 +25,10 @@ public abstract class BaseService<E, D, R, K> implements IBaseService<E, D, R, K
 
     @Override
     public R getById(K id) {
+        if (id == null) {
+            throw new IllegalArgumentException("ID cannot be null");
+        }
+
         return repository.findById(id)
                 .map(mapper::toResponseDto)
                 .orElseThrow(() -> new RuntimeException("Entity not found with id: " + id));
@@ -39,6 +43,10 @@ public abstract class BaseService<E, D, R, K> implements IBaseService<E, D, R, K
 
     @Override
     public R update(K id, D dto) {
+        if (id == null) {
+            throw new IllegalArgumentException("ID cannot be null");
+        }
+
         E entity = repository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Entity not found with id: " + id));
         updateEntity(entity, dto);
@@ -48,6 +56,13 @@ public abstract class BaseService<E, D, R, K> implements IBaseService<E, D, R, K
 
     @Override
     public void delete(K id) {
+        if (id == null) {
+            throw new IllegalArgumentException("ID cannot be null");
+        }
+        if (!repository.existsById(id)) {
+            throw new RuntimeException("Entity not found with id: " + id);
+        }
+
         repository.deleteById(id);
     }
